requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
dateparser==1.0.0
freezegun==1.1.0
idna==3.2
iniconfig==1.1.1
multidict==5.1.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websockets==9.1
yarl==1.6.3
